+++ packages/sources/nomics-test/src/endpoint/index.ts
@@ -1,0 +1,5 @@
+export { endpoint as filtered } from './filtered'
+export { endpoint as globalmarketcap } from './globalmarketcap'
+export { endpoint as crypto } from './crypto'
+export { endpoint as marketcap } from './marketcap'
+export { endpoint as volume } from './volume'