+++ packages/sources/cfbenchmarks-test/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @chainlink/cfbenchmarks-test-adapter
 
+## 1.1.1
+
+### Patch Changes
+
+- c152ed06e: Bumped v3 framework version
+
+## 1.1.0
+
+### Minor Changes
+
+- 70a7b9abd: Add missing rest transport in cfbenchmark v3 EA
+
 ## 1.0.5
 
 ### Patch Changes